# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ravyu Business Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "401":
          description: Invalid credentials
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "401":
          description: Invalid or expired refresh token
  /auth/firebase:
    post:
      operationId: firebaseAuth
      tags: [auth]
      summary: Sign in or register via Firebase Google token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FirebaseInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FirebaseAuthResult"
        "401":
          description: Invalid Firebase token

  # ── User ──────────────────────────────────────────────────────────────────
...
  /business/profile:
    get:
      operationId: getBusinessProfile
      tags: [business]
      summary: Get business profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
        "404":
          description: No profile yet
    post:
      operationId: createBusinessProfile
      tags: [business]
      summary: Create business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessProfileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
    put:
      operationId: updateBusinessProfile
      tags: [business]
      summary: Update business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessProfileInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"

  # ── Business Services ─────────────────────────────────────────────────────
  /business/services:
    get:
      operationId: getBusinessServices
      tags: [business]
      summary: List services
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessService"
    post:
      operationId: createBusinessService
      tags: [business]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessService"
  /business/services/{id}:
    put:
      operationId: updateBusinessService
      tags: [business]
      summary: Update a service
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessServiceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessService"
    delete:
      operationId: deleteBusinessService
      tags: [business]
      summary: Delete a service
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SignupInput:
      type: object
      required: [email, password, username]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        username:
          type: string
          minLength: 2
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RefreshInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    FirebaseInput:
      type: object
      required: [idToken]
      properties:
        idToken:
          type: string
    AuthTokens:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    FirebaseAuthResult:
      type: object
      required: [accessToken, refreshToken, user, isNewUser]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
        isNewUser:
          type: boolean

    # ── User ─────────────────────────────────────────────────────────────────
    UserProfile:
      type: object
      required: [id, email, username, profileComplete]
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        googleMapsUrl:
          type: ["string", "null"]
        businessType:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        challenges:
          type: ["array", "null"]
          items:
            type: string
        profileComplete:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    BusinessProfile:
      type: object
      required: [id, businessName, slug]
      properties:
        id:
          type: integer
        businessName:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zip:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        googleMapsUrl:
          type: ["string", "null"]
        googlePlaceId:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        secondaryColor:
          type: ["string", "null"]
        facebookUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        twitterUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        businessHours:
          type: ["string", "null"]
        pageViews:
          type: integer
        reviewClicks:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    BusinessProfileInput:
      type: object
      required: [businessName]
      properties:
        businessName:
          type: string
        description:
          type: string
        logoUrl:
          type: string
        coverImageUrl:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        googleMapsUrl:
          type: string
        googlePlaceId:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        facebookUrl:
          type: string
        instagramUrl:
          type: string
        twitterUrl:
          type: string
        linkedinUrl:
          type: string
        businessHours:
          type: string
    BusinessService:
      type: object
      required: [id, name, isActive, displayOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        isActive:
          type: boolean
        displayOrder:
          type: integer
    BusinessServiceInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: string
        duration:
          type: string
        isActive:
          type: boolean
        displayOrder:
          type: integer
...