# lib/api-zod/src/generated/types/businessService.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ravyu Business Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessService {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  duration?: string | null;
  isActive: boolean;
  displayOrder: number;
}